Proposal description fields
Temporary approach to Proposal description fields.
Description fields are the expandable details that help people drill
into the proposal. They are very specific to the type of proposal (e.g.
for
an LVG ballot measure, one of the fields might be "fiscal impact
statement").
For Tigard, I'm storing all these fields in
proposal.additional_description1
as a serialized JSON object of the following structure: