fixing redirection issue in IE9
Problem was that backbone creates a redirect when, for a non-pushstate
compliant browser like IE9, it changes to a hash-based url. This
interfered with the root redirection in the home_controller.

Solution is to just add a query string in the redirection that the home
controller can check for in order to know whether it should redirect
root or not.