don't continually save histogram state when unnecessary Fixes #50
The continual updates interfered with the operation of the sticky
decision board, which is based on a stable DOM (as a jquery plugin).

I would advocate making the sticky decision board a statebus driven
component, except that it appears future designs will not use the
sticky board; therefore it isn’t worth the time.